+++ src/config/settings.ts
@@ -136,6 +136,7 @@ const settings: Settings = {
     },
   },
   rpcSelectionStrategy: process.env.RPC_SELECTION_STRATEGY || RPCSelectionStrategies.BY_BLOCK_NUMBER,
+  feedsCacheRefreshCronRule: process.env.FEEDS_CACHE_REFRESH_CRON_RULE || '*/5 * * * *', // every five minutes
   feedsFile:
     process.env.FEEDS_FILE ||
     'https://raw.githubusercontent.com/umbrella-network/pegasus-feeds/main/prod/bsc/feeds.yaml',